Move update_code to release.py
